move models to a module, add config to get data from env, add webhook crc token endpoint
