+++ backend/app.py
@@ -1,3 +1,8 @@
+import base64
+import hashlib
+import hmac
+import os
+
 from typing import Optional
 
 
@@ -5,17 +10,10 @@ from fastapi import FastAPI
 
 app = FastAPI()
 
-from pydantic import BaseModel
+from backend.models import SignupBody, LoginBody
+from backend.config import Settings
 
-class SignupBody(BaseModel):
-    user_email: str
-    user_password: str
-    user_twitter_handle: str
-    full_name: Optional[str] = None
-
-class LoginBody(BaseModel):
-    user_password: str
-    user_twitter_handle: str
+settings = Settings()
 
 @app.get("/")
 async def root():
@@ -33,3 +31,14 @@ async def login(login_body: LoginBody):
 @app.get("/{user_id}/all_tweets", status_code=200)
 async def get_tweets(user_id: int):
     return {"id": user_id}
+
+
+@app.get("/webhook/crc", status_code=200)
+async def get_crc_response(crc_token: str):
+    # creates HMAC SHA-256 hash from incomming token and your consumer secret
+    # and constructs response data with base64 encoded hash
+    sha256_hash_digest = hmac.new(settings.twitter_secret, msg=crc_token, digestmod=hashlib.sha256).digest()
+    response = {
+        'response_token': 'sha256=' + base64.b64encode(sha256_hash_digest)
+    }
+    return response
