+++ backend/config.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class Settings(BaseSettings):
+    twitter_secret: str
+
+    class Config:
+        env_file = ".env"
