+++ backend/pyproject.toml
@@ -8,6 +8,8 @@ license = "MIT"
 [tool.poetry.dependencies]
 python = "^3.6"
 fastapi = "^0.63.0"
+pydantic = "^1.7.3"
+python-dotenv = "^0.15.0"
 
 [tool.poetry.dev-dependencies]
